Animación del dragon OK
dragon último nivel OK

+++ Assets/Scripts/AddHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class AddHealth : MonoBehaviour
 {
    
-    private float addHealth;
+    private int addHealth;
 
     private void Awake()
     {
