+++ Assets/Scripts/ScriptableObject Script/DrownerInfo.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu (fileName = "Datos Drowner", menuName = "ScriptableObjects/Drowner", order = 3)]
 public class DrownerInfo :  ScriptableObject
 {
-    public int maxHealth;
-    public int currentHealth;
-    public int attackDamage;
+    public float maxHealth;
+    public float currentHealth;
+    public float attackDamage;
 
 }
